requirements.txt
cached-property==1.5.2
click==8.0.4
cycler==0.10.0
h5py==3.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
matplotlib==2.1.1
nltk==3.6.7
numpy==1.16.4
opencv-python==4.7.0.72
Pillow==5.0.0
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
regex==2023.6.3
scapy==2.4.2
scipy==1.0.0
six==1.16.0
tqdm==4.64.1
typing_extensions==4.1.1
wget==3.2
zipp==3.6.0
